[REFACTOR] 인터페이스 작성 및 기능에 대한 주석 작성
